#8 minor naming refactoring
