+++ src/main/java/org/bearmug/aws/actions/Action.java
@@ -1,18 +1,18 @@
-package org.bearmug.aws;
+package org.bearmug.aws.actions;
 
 import org.telegram.telegrambots.api.methods.BotApiMethod;
 import org.telegram.telegrambots.api.methods.send.SendMessage;
 import org.telegram.telegrambots.api.objects.Update;
 
-public interface Router {
+public interface Action {
 
-    BotApiMethod answer();
+    BotApiMethod respond();
 
-    static Router route(Update update) {
+    static Action forInput(Update update) {
         return new DefaultRouter(update);
     }
 
-    class DefaultRouter implements Router {
+    class DefaultRouter implements Action {
         private final Update update;
 
         DefaultRouter(Update update) {
@@ -20,7 +20,7 @@ public interface Router {
         }
 
         @Override
-        public BotApiMethod answer() {
+        public BotApiMethod respond() {
             return new SendMessage(update.getMessage().getChatId(),
                     greetingsFor(update.getMessage().getChat().getFirstName()));
         }
